+++ Assets/Prefabs/Dialogues.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 19bf3a4d93a16664b80fbf9e62c232ce
-PrefabImporter:
+guid: a42d02506b241414ca79504bc344086b
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
