+++ Assets/Prefabs/Dialogues/DialoguePanel.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8a9da6bf15e9bfa44b364b7e75cb88f7
+guid: c22ee757a43186a439ea3a31ddccb9f9
 PrefabImporter:
   externalObjects: {}
   userData:
