+++ Assets/Scripts/Dialogues.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0be48f9d13d445448aff2831183ac942
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
