--- Assets/Scripts/TalkingNPC.cs
@@ -1,23 +1,0 @@
-﻿using UnityEngine;
-
-[RequireComponent(typeof(Dialogue))]
-public class TalkingNPC : MonoBehaviour
-{
-    private Dialogue dialogue;
-
-    // Start is called before the first frame update
-    private void Start()
-    {
-        dialogue = GetComponent<Dialogue>();
-    }
-
-    public void Talk()
-    {
-        dialogue.enabled = true;
-    }
-
-    public void CompleteTask()
-    {
-        dialogue.inkStory.variablesState["completed"] = true;
-    }
-}
